removed a lot of unused plotting code
